Add files via upload
This is a simple class assignment coded in assembly language in order to simulate an integer calculator. The 3 functions can be performed here are addition, subtraction, and multiplication. If you never coded in assembly before, hope this simple code gives you a little taste in assembly. Download it and try it out.